Add files via upload
There are my Flask learning files during these weeks, two weeks ago I have finished the flask2.py file for the purpose of learning how to get information about users' pcs. And after that, I was looking at some documents about flask redirecting behavior. And then, I make a web of getting the sum of two floats in the entered URL. It's funny. And on this weekend, after reading the chapters about temples and reviewing the grammar of HTML, I have made the very simple login page so as to practicing the template in Flask.